requirements.txt
nltk
streamlit
tenacity
openai
google-api-python-client
pytube
transformers
python-dotenv
google-auth-httplib2 
google-auth-oauthlib
langchain_google_genai
langchain
tiktoken